# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TartanHacks Backend",
    "version": "0.0.1"
  },
  "basePath": "/",
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token"
      }
    }
  },
  "paths": {
...
    "/check-in/user": {
      "put": {
        "summary": "Check a User In",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "tags": [
          "Check In Module"
        ],
        "description": "Checks in a user to a check in item. Access - Admin Users can check in any user for any check in item. Participants can only check themselves in for check-in items that have self check-in enabled.",
        "parameters": [
          {
            "in": "query",
            "name": "userID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "checkInItemID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Bad Request."
          },
          "404": {
            "description": "check in item or user not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
...
    "/requests/user/{userId}": {
      "get": {
        "summary": "View incoming and outgoing requests for a user",
        "tags": [
          "Team Requests Module"
        ],
        "description": "View incoming and outgoing requests for a user. Access - User",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "summary": "Query list of users",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "tags": [
          "Users Module"
        ],
        "description": "Retrieves list of all users in the database. Access - Admin",
        "responses": {
          "200": {
            "description": "Success."
          },
          "403": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Query user by user ID",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "tags": [
          "Users Module"
        ],
        "description": "Retrieves user info by user ID, Access - Admin or Owner",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "403": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    }
  },
...
}